% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorsph.R
\name{ll2cart}
\alias{ll2cart}
\title{Longitude latitude (and z) to Cartesian XYZ}
\usage{
ll2cart(llh, rad = 6378137, exag = 1)
}
\arguments{
\item{llh}{matrix of longitude, latitude, and optionally z (height above surface)}

\item{rad}{radius of sphere, assumed Earthish}

\item{exag}{exaggeration factor for height values}
}
\value{
matrix of XYZ coordinates (3 columns)
}
\description{
Longitude latitude (and z) to Cartesian XYZ
}
\examples{
ll2cart(cbind(c(-180, 0, 90, 180), c(-90, -42, 0, 45, 90)))
}
