% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_coords.R
\name{regular_coords}
\alias{regular_coords}
\title{Generate regular longitude latitude coordinates}
\usage{
regular_coords(rows = 24, force = FALSE, prob = 1)
}
\arguments{
\item{rows}{Number of rows of coordinates see Details}

\item{force}{set to \code{TRUE} to ignore restraint and create a lot of points}

\item{prob}{defaults to 1, otherwise set to a fraction to subsample the returned points}
}
\value{
matrix of longitude,latitude
}
\description{
Input is a number of \code{rows}, default is quite low.
}
\details{
\code{rows} argument is how many distinct latitude bands.

Code is taken from the croc package, which bundles logic from the Ocean Color
NASA website. https://github.com/sosoc/croc
}
\examples{
plot(regular_coords(10))


}
